requirements.txt
dnspython==1.16.0
hiredis==0.2.0
idna==2.6
redis==2.10.6
